requirements.txt
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.15
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.9.0
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.7.6.1
python-decouple==3.1
pytz==2018.7
virtualenv==16.7.2
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
